Extract pagination functions
