+++ src/actions/journal.php
@@ -4,20 +4,13 @@ namespace jbrowneuk;
 
 class Journal implements Action
 {
-    public static function modifier_pagination(array $pagination)
-    {
-        $totalPages = ceil($pagination['total_items'] / $pagination['items_per_page']);
-        $pages = range(1, $totalPages);
-        return $pages;
-    }
+    
 
     public function render(\PDO $pdo, PortfolioRenderer $renderer, array $pageParams)
     {
-        $page = $this->calculatePage($pageParams);
+        $page = parsePageNumber($pageParams);
         $posts = get_posts($pdo, $page);
         $pagination = get_post_pagination_data($pdo);
-
-        $renderer->registerPlugin(\Smarty\Smarty::PLUGIN_MODIFIER, 'pagination', '\jbrowneuk\Journal::modifier_pagination');
 
         $years = 2;
         $staleTimestamp = time() - (60 * 60 * 24 * 365 * $years);
@@ -27,17 +20,5 @@ class Journal implements Action
         $renderer->assign('pagination', ['page' => $page, ...$pagination]);
         $renderer->assign('staleTimestamp', $staleTimestamp);
         $renderer->displayPage('post-list');
-    }
-
-    // [TODO] convert to helper function so all pages can use it
-    private function calculatePage(array $pageParams) {
-        $index = array_find_key($pageParams, fn($item) => $item === 'page');
-
-        // Calculate whether the next param after 'page' is numeric
-        if (array_key_exists($index + 1, $pageParams) && is_numeric($pageParams[$index + 1])) {
-            return (int)$pageParams[$index + 1];
-        }
-
-        return 1;
     }
 }
