+++ src/core/pagination.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace jbrowneuk;
+
+/**
+ * Attempts to parse the page number from a provided set of parameters.
+ *
+ * Considering a query string split on slash, this function will try to find a
+ * pair in the form ['page', <number>] and return the number.
+ *
+ * @param array $pageParams an array of strings, usually generated by splitting
+ *              the REQUEST URI on slash
+ *
+ * @return int the parsed page number
+ */
+function parsePageNumber(array $pageParams)
+{
+    $index = array_find_key($pageParams, fn($item) => $item === 'page');
+
+    // Calculate whether the next param after 'page' is numeric
+    if (array_key_exists($index + 1, $pageParams) && is_numeric($pageParams[$index + 1])) {
+        return (int)$pageParams[$index + 1];
+    }
+
+    return 1;
+}
