+++ src/core/renderer.php
@@ -15,6 +15,13 @@ class PortfolioRenderer extends \Smarty\Smarty
         return self::$parsedown->text($input);
     }
 
+    public static function modifier_pagination(array $pagination)
+    {
+        $totalPages = ceil($pagination['total_items'] / $pagination['items_per_page']);
+        $pages = range(1, $totalPages);
+        return $pages;
+    }
+
     public function __construct()
     {
         parent::__construct();
@@ -23,6 +30,7 @@ class PortfolioRenderer extends \Smarty\Smarty
         $this->setCacheDir('smarty/cache');
 
         $this->registerPlugin(\Smarty\Smarty::PLUGIN_MODIFIER, 'parsedown', '\jbrowneuk\PortfolioRenderer::modifier_parsedown');
+        $this->registerPlugin(\Smarty\Smarty::PLUGIN_MODIFIER, 'pagination', '\jbrowneuk\PortfolioRenderer::modifier_pagination');
     }
 
     public function setStyleRoot(string $directory)
