+++ src/index.php
@@ -5,6 +5,7 @@ namespace jbrowneuk;
 require_once '../vendor/autoload.php';
 
 require_once './core/action.php';
+require_once './core/pagination.php';
 require_once './core/renderer.php';
 
 require_once './database/album.php';
